delete confirmation + grey button tooltip

+++ src/components/SingleList.jsx
@@ -2,9 +2,28 @@ import { deleteList } from '@/api';
 import { getAuth } from 'firebase/auth';
 import { useEffect, useState } from 'react';
 import { FaShareNodes } from 'react-icons/fa6';
-import { RiDeleteBin5Fill } from 'react-icons/ri';
 import { useNavigate } from 'react-router-dom';
 import toast from 'react-hot-toast';
+import {
+	AlertDialog,
+	AlertDialogAction,
+	AlertDialogCancel,
+	AlertDialogContent,
+	AlertDialogDescription,
+	AlertDialogFooter,
+	AlertDialogHeader,
+	AlertDialogTitle,
+	AlertDialogTrigger,
+} from './ui/alert-dialog';
+import { Button } from './ui/button';
+import { Trash2 } from 'lucide-react';
+import {
+	Tooltip,
+	TooltipContent,
+	TooltipProvider,
+	TooltipTrigger,
+} from '@/components/ui/tooltip';
+
 export function SingleList({
 	name,
 	path,
@@ -13,6 +32,7 @@ export function SingleList({
 	handleShareModalClick,
 	setSelectedItem,
 }) {
+	const [isAlertOpen, setIsAlertOpen] = useState(false);
 	const [collectionId, setCollectionId] = useState('');
 	const singleListPath = path.split('/')[0];
 	const email = getAuth().currentUser.email;
@@ -43,6 +63,7 @@ export function SingleList({
 			setListPath('');
 		}
 		toast.success(`List ${name} was deleted`);
+		setIsAlertOpen(false);
 	};
 
 	return (
@@ -61,16 +82,59 @@ export function SingleList({
 					<FaShareNodes />
 				</button>
 				{getAuth().currentUser.uid === singleListPath ? (
-					<button
-						onClick={() => handleDeleteClick(name)}
-						className="text-pink-500  hover:text-pink-600"
-					>
-						<RiDeleteBin5Fill />
-					</button>
+					<AlertDialog open={isAlertOpen} onOpenChange={setIsAlertOpen}>
+						<AlertDialogTrigger asChild>
+							<Button
+								className="bg-transparent hover:bg-transparent"
+								type="button"
+								onClick={() => setIsAlertOpen(true)}
+							>
+								<Trash2 className="text-pink-500  hover:text-pink-600" />
+							</Button>
+						</AlertDialogTrigger>
+						<AlertDialogContent>
+							<AlertDialogHeader>
+								<AlertDialogTitle className="text-pink-500">
+									Are you absolutely sure?
+								</AlertDialogTitle>
+								<AlertDialogDescription>
+									This will permanently delete your list. Do you really want to
+									delete {name}?
+								</AlertDialogDescription>
+							</AlertDialogHeader>
+							<AlertDialogFooter>
+								<AlertDialogCancel
+									className="p-5 bg-red-600 rounded text-white hover:bg-red-800"
+									onClick={() => setIsAlertOpen(false)}
+								>
+									Cancel
+								</AlertDialogCancel>
+								<AlertDialogAction
+									className="p-5 bg-green-600 rounded text-white hover:bg-green-800"
+									onClick={() => handleDeleteClick(name)}
+								>
+									Continue
+								</AlertDialogAction>
+							</AlertDialogFooter>
+						</AlertDialogContent>
+					</AlertDialog>
 				) : (
-					<div className="text-gray-500">
-						<RiDeleteBin5Fill />
-					</div>
+					<TooltipProvider>
+						<Tooltip delayDuration={100}>
+							<TooltipTrigger asChild>
+								<Button
+									className="bg-transparent hover:bg-transparent"
+									type="button"
+									onClick={() => setIsAlertOpen(true)}
+								>
+									<Trash2 className="text-gray-500" />
+								</Button>
+							</TooltipTrigger>
+							<TooltipContent>
+								<p>You cannot delete a list you don&#39;t own!</p>
+							</TooltipContent>
+						</Tooltip>
+					</TooltipProvider>
 				)}
 			</div>
 		</li>
